fix low-volume bug
(this caused the glitchy sounds after looping crooner.d00)

learned the hard way that x >> 32 is the same as x >> 0
in fact x >> y === x >> (y&31).

so for any log-volumes >= 256*31 we just produce 0. hate to add a branch
in the inner loops just for this but it's a very predictable branch at
least.

also, buffer allocations in the audio callback are only done once.